==> 01_fetch_data.ipynb <==
from pathlib import Path
import pandas as pd
import requests

def fetch_raw_data(year: int, month: int) -> str:
    url = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02}.parquet"
    response = requests.get(url)
    if response.status_code == 200:
        # Use pathlib to construct the path
        path = Path("..") / "data" / "raw" / f"rides_{year}_{month:02}.parquet"
        # Ensure the parent directory exists
        path.parent.mkdir(parents=True, exist_ok=True)
        # Write the content to the file
        path.write_bytes(response.content)
        print(f"Successfully fetched: {str(path)}")
        return str(path)
    else:
        raise Exception(f"{url} is not available")


fetch_raw_data(2023, 1)

# from pathlib import Path

# # Filter and rename columns
# rides = rides[final_filter]
# rides = rides[["tpep_pickup_datetime", "PULocationID"]]
# rides.rename(columns={
#     "tpep_pickup_datetime": "pickup_datetime",
#     "PULocationID": "pickup_location_id"
# }, inplace=True)

# # Define year and month
# year = 2023
# month = 1

# # Define the directory and ensure it exists
# path = Path("..") / "data" / "processed"
# path.mkdir(parents=True, exist_ok=True)

# # Define file path and save as Parquet
# file_path = path / f"rides_{year}_{month:02}.parquet"
# rides.to_parquet(file_path, engine="pyarrow", index=False)
