==> 05_transform_raw_data_into_features_and_targets.ipynb <==
## Steps
1. Download all raw data for a given year and month
1. For each raw data, apply the filter before saving it
1. Transform the saved raw data into TS data
1. Convert the ts data into features and targets
1. Save the transformed data


Main objective is to write utility functions to do all these things so we can reuse them later. 

%load_ext autoreload
%autoreload 2

import sys
import os

# Define the absolute path to the 'src' directory
src_path = "/Users/manu/Documents/SPEACIAL_TOPICS/src"

# Append it to sys.path
sys.path.append(src_path)

# Now import the module
from data_utils import load_and_process_taxi_data


rides = load_and_process_taxi_data(year=2023)

rides

from src.data_utils import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data.head()

ts_data.shape

from src.data_utils import transform_ts_data_info_features_and_target_loop

features, targets = transform_ts_data_info_features_and_target_loop(ts_data, window_size=24*28, step_size=24)


features

import numpy as np
window_size = 4
step_size = 1
num_windows = 4

step_size * np.arange(num_windows)[:, None]  # Output: array([[0], [1], [2], [3]])

indices = np.arange(window_size)[None, :] + step_size * np.arange(num_windows)[:, None]


window_size = 4
step_size = 1
num_windows = 4

import numpy as np
np.arange(num_windows)

np.arange(num_windows)[:, None]

step_size * np.arange(num_windows)[:, None]

indices = np.arange(window_size)[None, :] + step_size * np.arange(num_windows)[:, None]


np.arange(window_size)[None, :]

np.arange(window_size)[None, :] + step_size * np.arange(num_windows)[:, None]

# Array 1: A 2D array with 2 rows and 2 columns
array1 = np.array([[1, 2],
                   [3, 4]])

# Array 2: A 2D array with 2 rows and 1 column
array2 = np.array([[5],
                   [6]])

# Array 3: A 2D array with 2 rows and 1 column
array3 = np.array([[7],
                   [8]])


np.hstack([array1, array2, array3])

from src.data_utils import transform_ts_data_info_features_and_target_loop

features, targets = transform_ts_data_info_features_and_target_loop(ts_data, window_size=24*28*1, step_size=24)


tabular_data = features
tabular_data["target"] = targets

from src.config import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet", engine="pyarrow")

features, targets = transform_ts_data_info_features_and_target_loop(ts_data[ts_data["pickup_location_id"]==43], window_size=12, step_size=1)


features

ts_data[ts_data["pickup_location_id"]==43].head(36)

from src.data_utils import transform_ts_data_info_features_and_target

features, targets = transform_ts_data_info_features_and_target(ts_data[ts_data["pickup_location_id"]==43], window_size=4, step_size=1)


features

from src.data_utils import transform_ts_data_info_features_and_target

features, targets = transform_ts_data_info_features_and_target(ts_data[ts_data["pickup_location_id"]==43], window_size=4, step_size=1)


print(features.columns)

features[["rides_t-4", "rides_t-3"]]
